Update, now the configuration contains the language switcher, and the file `file_manager_language.php` is removed
